rename utils to primitiveextension
