+++ cmd/warc/cmd/ls/ls.go
@@ -23,7 +23,7 @@ import (
 	"sort"
 	"strconv"
 
-	"github.com/nlnwa/gowarc/pkg/utils"
+	ext "github.com/nlnwa/gowarc/pkg/primitiveextension"
 	"github.com/nlnwa/gowarc/warcoptions"
 	"github.com/nlnwa/gowarc/warcreader"
 	"github.com/nlnwa/gowarc/warcrecord"
@@ -93,7 +93,7 @@ func readFile(c *conf, fileName string) {
 			break
 		}
 		if len(c.id) > 0 {
-			if !utils.Contains(c.id, wr.WarcHeader().Get(warcrecord.WarcRecordID)) {
+			if !ext.Contains(c.id, wr.WarcHeader().Get(warcrecord.WarcRecordID)) {
 				continue
 			}
 		}
@@ -110,6 +110,6 @@ func readFile(c *conf, fileName string) {
 
 func printRecord(offset int64, record warcrecord.WarcRecord) {
 	recordID := record.WarcHeader().Get(warcrecord.WarcRecordID)
-	targetURI := utils.CropString(record.WarcHeader().Get(warcrecord.WarcTargetURI), 100)
+	targetURI := ext.CropString(record.WarcHeader().Get(warcrecord.WarcTargetURI), 100)
 	fmt.Printf("%v\t%s\t%s \t%s\n", offset, recordID, record.Type(), targetURI)
 }
